Issue2_vectorDB/requirements.txt
torch
torchvision
pillow
numpy
sqlalchemy
timm
psycopg2-binary
scikit-learn
pgvector